Update Work Item component titles
